requirements/base.txt
fastapi==0.100.1
tortoise-orm
aerich
asyncpg
pydantic-settings